Columns of PinnedTags can have Include additional tags for this column enabled
This adds not only filter capabilities, which are not relevent to this project
like `not` and `all` but also has the `any` field, which is an array where
you may add more tags.

Unfortunatly, I was unable to create a precise select query for this, as the
database field `data` is not jsonb. I think parsing json  is really the most
optimal way under these circumstances.